requirements.txt
Fabric==1.8.2
Flask==1.0.2
Flask-SQLAlchemy==1.0
gunicorn==18.0
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.3
Werkzeug==0.9.4
coverage==3.7.1
ecdsa==0.10
itsdangerous==0.23
paramiko==1.12.2
pycrypto==2.6.1
wsgiref==0.1.2
Flask-WTF==0.9.4
pytest-xdist==1.10
py-bcrypt==0.4
py-jwt==2.0.1